paper-analysis-group-assessment-main/paper-analysis-group-assessment-main/requirements.txt
amqp==5.1.1
asgiref==3.5.2
auto-labeling-pipeline==0.1.21
backports.zoneinfo==0.2.1
billiard==3.6.4.0
blis==0.7.8
boto3==1.24.27
botocore==1.27.27
cachetools==5.2.0
catalogue==2.0.7
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cymem==2.0.6
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==4.0.6
django-celery-results==2.2.0
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-drf-filepond==0.4.1
django-filter==21.1
django-health-check==3.16.5
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.9
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-xml==2.0.0
doccano==1.8.0
drf-yasg==1.20.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.3.0/en_core_web_trf-3.3.0-py3-none-any.whl
environs==9.5.0
et-xmlfile==1.1.0
filelock==3.7.1
filetype==1.1.0
fire==0.4.0
fonttools==4.34.4
furl==2.1.3
google-api-core==2.8.2
google-auth==2.9.0
google-cloud-core==2.3.1
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
greenlet==1.1.2
gunicorn==20.1.0
huggingface-hub==0.8.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
kombu==5.2.4
langcodes==3.3.0
lml==0.1.0
lxml==4.9.1
MarkupSafe==2.1.1
marshmallow==3.17.0
more-itertools==7.0.0
murmurhash==1.0.7
numpy==1.23.0
opencv-python==4.6.0.66
openpyxl==3.0.10
orderedmultidict==1.0.1
packaging==21.3
pandas==1.4.3
pathy==0.6.2
pdf2docx==0.5.5
pdfminer.six==20211012
preshed==3.0.6
prompt-toolkit==3.0.30
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-xlsx==0.6.0
Pygments==2.12.0
PyMuPDF==1.20.1
pyparsing==3.0.9
PyPDF2==2.4.2
python-dateutil==2.8.2
python-docx==0.8.10
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.28.1
rich==10.16.2
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
scikit-learn==1.1.1
scipy==1.8.1
seqeval==1.2.2
shortuuid==1.0.9
simplify-docx==0.1.1
six==1.16.0
smart-open==5.2.1
spacy==3.3.1
spacy-alignments==0.8.5
spacy-legacy==3.0.9
spacy-loggers==1.0.2
spacy-transformers==1.1.7
spacypdfreader==0.2.1
SQLAlchemy==1.4.39
sqlparse==0.4.2
srsly==2.4.3
termcolor==1.1.0
texttable==1.6.4
thinc==8.0.17
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.0
tqdm==4.64.0
transformers==4.20.1
typer==0.4.2
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
waitress==2.1.2
wasabi==0.9.1
wcwidth==0.2.5
whitenoise==6.2.0
wincertstore==0.2
pdfplumber==0.11.6
python-docx==0.8.10
pymongo==4.12.0
